requirements.txt
numpy==1.26.0
matplotlib==3.8.0
pandas==2.1.1
seaborn==0.13.0
flask==3.0.0
Flask-Cors==4.0.0
opencv-python==4.8.1.78
tensorflow==2.8
tensorflow-addons==0.22.0
scikit-learn==1.3.1
mediapipe==0.10.7
plotly==5.17.0
PyYAML==6.0.1
networkx==3.2.1
scikit-learn==1.3.1
Pillow==10.1.0